Replace print statements with logging calls on mock server
