+++ requirements.txt
@@ -11,6 +11,8 @@ mypy
 brotli
 brotli-asgi
 awslambdaric-stubs
+requests
+shutil
 testing.postgresql
 testing.redis
 aioboto3
