+++ tests/mock_server.py
@@ -1,9 +1,11 @@
-import pytest
-import requests
-import shutil
 import signal
 import subprocess as sp
 import time
+import logging
+
+import pytest
+import requests
+import shutil
 
 _proxy_bypass = {
     "http": None,
@@ -15,16 +17,16 @@ def start_service(service_name, host, port):
     moto_svr_path = shutil.which("moto_server")
     args = [moto_svr_path, service_name, "-H", host, "-p", str(port)]
     process = sp.Popen(args, stdin=sp.PIPE, stdout=sp.PIPE, stderr=sp.PIPE)
-    url = "http://{host}:{port}".format(host=host, port=port)
+    url = f"http://{host}:{port}"
 
     for i in range(0, 30):
         output = process.poll()
         if output is not None:
-            print("moto_server exited status {0}".format(output))
+            logging.info(f"moto_server exited status {output}")
             stdout, stderr = process.communicate()
-            print("moto_server stdout: {0}".format(stdout))
-            print("moto_server stderr: {0}".format(stderr))
-            pytest.fail("Can not start service: {}".format(service_name))
+            logging.info(f"moto_server stdout: {stdout}")
+            logging.info(f"moto_server stderr: {stderr}")
+            pytest.fail(f"Can not start service: {service_name}")
 
         try:
             # we need to bypass the proxies due to monkeypatches
@@ -34,7 +36,7 @@ def start_service(service_name, host, port):
             time.sleep(0.5)
     else:
         stop_process(process)  # pytest.fail doesn't call stop_process
-        pytest.fail("Can not start service: {}".format(service_name))
+        pytest.fail(f"Can not start service: {service_name}")
 
     return process
 
@@ -47,27 +49,5 @@ def stop_process(process):
         process.kill()
         outs, errors = process.communicate(timeout=20)
         exit_code = process.returncode
-        msg = "Child process finished {} not in clean way: {} {}".format(
-            exit_code, outs, errors
-        )
+        msg = f"Child process finished {exit_code} not in clean way: {outs} {errors}"
         raise RuntimeError(msg)
-
-
-@pytest.fixture(scope="session")
-def dynamodb2_server(aws_credentials):
-    host = "localhost"
-    port = 5001
-    url = "http://{host}:{port}".format(host=host, port=port)
-    process = start_service("dynamodb2", host, port)
-    yield url
-    stop_process(process)
-
-
-@pytest.fixture(scope="session")
-def s3_server(aws_credentials):
-    host = "localhost"
-    port = 5002
-    url = "http://{host}:{port}".format(host=host, port=port)
-    process = start_service("s3", host, port)
-    yield url
-    stop_process(process)
